chore: rename tw merge wrapper
