+++ web/src/components/Input.tsx
@@ -1,5 +1,5 @@
 import { HTMLProps } from "react";
-import { cm } from "../utils/tailwind-merge";
+import { cm } from "../utils/tailwindMerge";
 
 export type CommonInputProps = {
   containerClassName?: string;
